pyproject.toml
[tool.poetry]
name = "pinnacle"
version = "0.1.0"
description = ""
authors = ["nt54hamnghi <hamnghi250699@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9.7"
attrs = "^21.4.0"
httpx = "^0.23.0"
python-dotenv = "0.16.0"
psutil = "^5.9.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^22.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
